certstream-monitor/requirements.txt
certstream==1.12
tld==0.7.9
python_Levenshtein==0.12.0
websocket-client==0.58.0
PyYAML==5.1
psycopg2==2.9.9